# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Calendros API Documentation",
        "description": "API documentation for Calendros - A comprehensive calendar management system",
        "contact": {
            "email": "support@calendros.example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://http://calendros.test/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://prm.positive.io.vn/api",
            "description": "Production Server"
        },
        {
            "url": "https://calendros.montserrat.id.vn/api",
            "description": "Test Server"
        }
    ],
    "paths": {
...
        "/calendars/{calendar_id}/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "List events in a calendar",
                "description": "Get all events for a specific calendar with optional date range filtering",
                "operationId": "f16c4fb9230fbca130cb1dc1cb38e9f6",
                "parameters": [
                    {
                        "name": "calendar_id",
                        "in": "path",
                        "description": "Calendar ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "filter[start_time][gte]",
                        "in": "query",
                        "description": "Filter events starting from this date (ISO 8601 format)",
                        "schema": {
                            "type": "string",
                            "format": "datetime",
                            "example": "2025-11-01T00:00:00Z"
                        }
                    },
                    {
                        "name": "filter[end_time][lte]",
                        "in": "query",
                        "description": "Filter events ending before this date (ISO 8601 format)",
                        "schema": {
                            "type": "string",
                            "format": "datetime",
                            "example": "2025-11-30T23:59:59Z"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of events",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Team Meeting"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Weekly sync"
                                                    },
                                                    "start_time": {
                                                        "type": "string",
                                                        "format": "datetime",
                                                        "example": "2025-11-15T10:00:00Z"
                                                    },
                                                    "end_time": {
                                                        "type": "string",
                                                        "format": "datetime",
                                                        "example": "2025-11-15T11:00:00Z"
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": "Conference Room A"
                                                    },
                                                    "is_all_day": {
                                                        "type": "boolean",
                                                        "example": false
                                                    },
                                                    "calendar_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Calendar not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Create a new event",
                "description": "Create a new event in a specific calendar",
                "operationId": "d6dfedb9bde45b1145c541b6f9acad36",
                "parameters": [
                    {
                        "name": "calendar_id",
                        "in": "path",
                        "description": "Calendar ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Product Launch"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Launch event details"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2025-12-01T09:00:00Z"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2025-12-01T17:00:00Z"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Main Auditorium"
                                    },
                                    "is_all_day": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Event created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Product Launch"
                                                },
                                                "calendar_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}